+++ bpm-server/persistence/src/main/java/br/com/bpm/persistence/features/solicitation/SolicitationRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SolicitationRepository  extends JpaRepository<Solicitation, Long> {
+public interface SolicitationRepository  extends JpaRepository<Solicitation, Integer> {
 
 }
